requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
pydantic==2.9.2
python-dateutil==2.9.0.post0
feedparser==6.0.11
requests==2.32.3
readability-lxml==0.8.1
lxml[html_clean]==5.3.0
beautifulsoup4==4.12.3
pytz==2024.2
APScheduler==3.10.4
Jinja2==3.1.4
httpx==0.27.2
orjson==3.10.7
